Pulled this into a legit "game" ready for demo
Stuck with the 2 player mode for now, got a lot of the unit stats going,
controls work the way envisioned for the most part. Actual level
progression. Can only use one unit per turn now. Other adjustments. This
would actually be in pretty solid form if we had stats displayed below.
That should be a high priority to get this into top condition. Win and
lose screens need to be developed a bit more too.