Touched on audio and worked on interface
See what I did with the channels in sfx.lua. From here on out I'll only
be working on the AI and splitting Level into both the single player/AI
mode and the 2 player mode. That will come after the single player mode
with the AI is complete. For our submission we're only going to have one
AI type and it's going to be rather dumb and predictable. C'est la vie.